Move multimap primitive into the io.atomix.core.map package for consistency with other primitives.
