+++ core/src/main/java/io/atomix/core/Atomix.java
@@ -28,8 +28,8 @@ import io.atomix.core.impl.CorePrimitivesService;
 import io.atomix.core.lock.DistributedLock;
 import io.atomix.core.map.AtomicCounterMap;
 import io.atomix.core.map.ConsistentMap;
+import io.atomix.core.map.ConsistentMultimap;
 import io.atomix.core.map.ConsistentTreeMap;
-import io.atomix.core.multimap.ConsistentMultimap;
 import io.atomix.core.profile.Profile;
 import io.atomix.core.queue.WorkQueue;
 import io.atomix.core.semaphore.DistributedSemaphore;
