+++ core/src/main/java/io/atomix/core/impl/CorePrimitivesService.java
@@ -37,10 +37,10 @@ import io.atomix.core.map.AtomicCounterMap;
 import io.atomix.core.map.AtomicCounterMapType;
 import io.atomix.core.map.ConsistentMap;
 import io.atomix.core.map.ConsistentMapType;
+import io.atomix.core.map.ConsistentMultimap;
+import io.atomix.core.map.ConsistentMultimapType;
 import io.atomix.core.map.ConsistentTreeMap;
 import io.atomix.core.map.ConsistentTreeMapType;
-import io.atomix.core.multimap.ConsistentMultimap;
-import io.atomix.core.multimap.ConsistentMultimapType;
 import io.atomix.core.queue.WorkQueue;
 import io.atomix.core.queue.WorkQueueType;
 import io.atomix.core.semaphore.DistributedSemaphore;
