+++ core/src/main/java/io/atomix/core/map/MultimapEvent.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.core.multimap;
+package io.atomix.core.map;
 
 import com.google.common.base.MoreObjects;
