+++ core/src/main/java/io/atomix/core/map/MultimapEventListener.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2018-present Open Networking Foundation
+ * Copyright 2017-present Open Networking Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,19 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.core.multimap.impl;
-
-import io.atomix.primitive.protocol.PrimitiveProtocol;
-import io.atomix.protocols.raft.MultiRaftProtocol;
+package io.atomix.core.map;
 
 /**
- * Raft multimap test.
+ * Listener to be notified about updates to a ConsistentMultimap.
  */
-public class RaftConsistentSetMultimapTest extends ConsistentSetMultimapTest {
-  @Override
-  protected PrimitiveProtocol protocol() {
-    return MultiRaftProtocol.builder()
-        .withMaxRetries(5)
-        .build();
-  }
+public interface MultimapEventListener<K, V> {
+
+  /**
+   * Reacts to the specified event.
+   *
+   * @param event the event
+   */
+  void event(MultimapEvent<K, V> event);
 }
