+++ core/src/main/java/io/atomix/core/map/impl/ConsistentMultimapFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-present Open Networking Foundation
+ * Copyright 2017-present Open Networking Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,8 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package io.atomix.core.map.impl;
 
 /**
- * Distributed multimap primitive interfaces.
+ * Consistent multimap factory.
  */
-package io.atomix.core.multimap;
+public interface ConsistentMultimapFactory {
+}
