+++ core/src/main/java/io/atomix/core/map/impl/ConsistentSetMultimapService.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.core.multimap.impl;
+package io.atomix.core.map.impl;
 
 import io.atomix.primitive.operation.Command;
 import io.atomix.primitive.operation.Query;
