+++ core/src/main/java/io/atomix/core/map/impl/DelegatingAsyncConsistentMultimap.java
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package io.atomix.core.multimap.impl;
+package io.atomix.core.map.impl;
 
 import io.atomix.core.collection.AsyncDistributedCollection;
-import io.atomix.core.multimap.AsyncConsistentMultimap;
-import io.atomix.core.multimap.ConsistentMultimap;
-import io.atomix.core.multimap.MultimapEventListener;
+import io.atomix.core.map.AsyncConsistentMultimap;
+import io.atomix.core.map.ConsistentMultimap;
+import io.atomix.core.map.MultimapEventListener;
 import io.atomix.core.set.AsyncDistributedMultiset;
 import io.atomix.core.set.AsyncDistributedSet;
 import io.atomix.primitive.DelegatingAsyncPrimitive;
