+++ core/src/main/java/io/atomix/core/map/impl/TranscodingAsyncConsistentMultimap.java
@@ -14,17 +14,17 @@
  * limitations under the License.
  */
 
-package io.atomix.core.multimap.impl;
+package io.atomix.core.map.impl;
 
 import com.google.common.collect.ImmutableMultiset;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Multiset;
 import io.atomix.core.collection.AsyncDistributedCollection;
 import io.atomix.core.collection.impl.TranscodingAsyncDistributedCollection;
-import io.atomix.core.multimap.AsyncConsistentMultimap;
-import io.atomix.core.multimap.ConsistentMultimap;
-import io.atomix.core.multimap.MultimapEvent;
-import io.atomix.core.multimap.MultimapEventListener;
+import io.atomix.core.map.AsyncConsistentMultimap;
+import io.atomix.core.map.ConsistentMultimap;
+import io.atomix.core.map.MultimapEvent;
+import io.atomix.core.map.MultimapEventListener;
 import io.atomix.core.set.AsyncDistributedMultiset;
 import io.atomix.core.set.AsyncDistributedSet;
 import io.atomix.core.set.impl.TranscodingAsyncDistributedMultiset;
